requirements.txt
boto3
openai
python-hcl2

